refactor: App클래스에서 경주와 관련된 메서드들은 Race클래스로 분리
